Issue #327 Migrated MoveiBrowser to MVP
Initial cut is done. Categories is working again.
